requirements.txt
colorama==0.4.6
python-dotenv==1.1.0
openai-agents==0.6.2
docker==7.0.0
openai>=1.0.0
fastapi==0.115.9
uvicorn[standard]==0.34.0
tiktoken>=0.5.0

